pyproject.toml
[project]
name = "placeholderproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sympy>=1.14.0",
    "tavily-python>=0.7.12",
    "transformers>=4.56.1",
]

[tool.uv.workspace]
members = [
    "services/rag",
]
